Move the standalone taso test too
